feat(api): add missing return if file has no submission location in downloadFile, add submission location !=  check when deleting file from inbox

+++ sda/cmd/api/api.go
@@ -472,6 +472,13 @@ func deleteFile(c *gin.Context) {
 		return
 	}
 
+	if location == "" {
+		log.Errorf("fileID: %s has no known submission location", fileID)
+		c.AbortWithStatusJSON(http.StatusInternalServerError, "failed to find file location")
+
+		return
+	}
+
 	filePath = helper.UnanonymizeFilepath(filePath, submissionUser)
 	for count := 1; count <= 5; count++ {
 		err = inboxWriter.RemoveFile(c, location, filePath)
@@ -560,8 +567,9 @@ func downloadFile(c *gin.Context) {
 	}
 	if location == "" {
 		log.Errorf("fileID: %s has no known submission location", fileID)
-
 		c.AbortWithStatusJSON(http.StatusInternalServerError, "failed to find file location")
+
+		return
 	}
 
 	// Get inbox file handle #noqa
